updateJira to TS; cleanup
